+++ src/domains/projects/chapter-assignments/project-chapter-assignments.route.ts
@@ -9,7 +9,7 @@ import { requireProjectAccess } from '@/domains/projects/project-auth.middleware
 import { PROJECT_ACTIONS } from '@/domains/projects/projects.types';
 import { PERMISSIONS } from '@/lib/permissions';
 import { getHttpStatus } from '@/lib/types';
-import { authenticateUser, requirePermission, requireSelf } from '@/middlewares/role-auth';
+import { authenticateUser, requirePermission } from '@/middlewares/role-auth';
 import { server } from '@/server/server';
 
 import * as service from './project-chapter-assignments.service';
@@ -19,7 +19,6 @@ import {
   assignUserInputSchema,
   chapterAssignmentProgressResponseSchema,
   chapterAssignmentResponseSchema,
-  memberChapterAssignmentsResponseSchema,
 } from './project-chapter-assignments.types';
 
 const projectIdParam = z.object({ projectId: z.coerce.number().int().positive() });
@@ -294,90 +293,5 @@ server.openapi(assignSelectedRoute, async (c) => {
 
   const result = await service.assignSelectedChapters(projectId, assignments);
   if (result.ok) return c.json(result.data, HttpStatusCodes.OK);
-  return c.json({ message: result.error.message }, getHttpStatus(result.error) as never);
-});
-
-const getUserChapterAssignmentsRoute = createRoute({
-  tags: ['Projects - Chapter Assignments'],
-  method: 'get',
-  path: '/project/member/{userId}/chapter-assignments',
-  middleware: [
-    authenticateUser,
-    requirePermission(PERMISSIONS.PROJECT_VIEW),
-    requireSelf(),
-  ] as const,
-  summary: 'Get all chapter assignments for a user across all their projects',
-  description:
-    'Fetches all projects the user belongs to, then returns all chapter assignments ' +
-    'across those projects in a single flat list — including unassigned ones.',
-  request: {
-    params: z.object({
-      userId: z.coerce
-        .number()
-        .int()
-        .positive()
-        .openapi({
-          param: { name: 'userId', in: 'path', required: true },
-          example: 1,
-        }),
-    }),
-    query: z.object({
-      excludeProjectIds: z
-        .string()
-        .optional()
-        .transform((val) => val?.split(',').map(Number).filter(Boolean) ?? [])
-        .openapi({
-          param: { name: 'excludeProjectIds', in: 'query', required: false },
-          description: 'Comma-separated list of project IDs to exclude',
-          example: '97,98',
-        }),
-      updatedAfter: z
-        .string()
-        .optional()
-        .transform((val) => (val ? new Date(val) : undefined))
-        .pipe(z.date().optional())
-        .openapi({
-          param: { name: 'updatedAfter', in: 'query', required: false },
-          description: 'Return only assignments updated after this ISO timestamp',
-          example: '2025-01-01T00:00:00.000Z',
-        }),
-    }),
-  },
-  responses: {
-    [HttpStatusCodes.OK]: jsonContent(
-      memberChapterAssignmentsResponseSchema.openapi('UserChapterAssignments'),
-      'Flat list of all chapter assignments across all user projects'
-    ),
-    [HttpStatusCodes.UNAUTHORIZED]: jsonContent(
-      createMessageObjectSchema('Unauthorized'),
-      'Authentication required'
-    ),
-    [HttpStatusCodes.FORBIDDEN]: jsonContent(
-      createMessageObjectSchema('Forbidden'),
-      'Access denied'
-    ),
-    [HttpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
-      createMessageObjectSchema(HttpStatusPhrases.INTERNAL_SERVER_ERROR),
-      'Internal server error'
-    ),
-  },
-});
-server.openapi(getUserChapterAssignmentsRoute, async (c) => {
-  const { userId } = c.req.valid('param');
-  const { excludeProjectIds, updatedAfter } = c.req.valid('query');
-
-  const result = await service.getChapterAssignmentsByUserId(
-    userId,
-    excludeProjectIds,
-    updatedAfter
-  );
-  if (result.ok)
-    return c.json(
-      {
-        syncedAt: new Date().toISOString(),
-        data: result.data,
-      },
-      HttpStatusCodes.OK
-    );
   return c.json({ message: result.error.message }, getHttpStatus(result.error) as never);
 });
