+++ src/domains/projects/chapter-assignments/project-chapter-assignments.types.ts
@@ -32,27 +32,6 @@ export const chapterAssignmentProgressResponseSchema = z.object({
   updatedAt: z.date().nullable(),
 });
 
-export const memberChapterAssignmentResponseSchema = z.object({
-  chapterAssignmentId: z.number().int(),
-  projectId: z.number().int(),
-  projectUnitId: z.number().int(),
-  bibleId: z.number().int(),
-  bookId: z.number().int(),
-  chapterNumber: z.number().int(),
-  assignedUserId: z.number().int().nullable(),
-  peerCheckerId: z.number().int().nullable(),
-  status: z.string(),
-  submittedTime: z.string().nullable(),
-  createdAt: z.string().nullable(),
-  updatedAt: z.string().nullable(),
-});
-
-export const memberChapterAssignmentsResponseSchema = z.object({
-  syncedAt: z.string(),
-  data: memberChapterAssignmentResponseSchema.array(),
-});
-
-export type MemberChapterAssignmentResponse = z.infer<typeof memberChapterAssignmentResponseSchema>;
 // ─── Assign-all input ─────────────────────────────────────────────────────────
 
 export const assignUserInputSchema = z
